-- ACBrLCDPR [-] Bloco 0040 e Bloco 0045 campo "Participação" é do tipo Double. Por diogo_maxx. Veja: https://www.projetoacbr.com.br/forum/topic/54594-acbrlcdpr-vers%C3%A3o-12-contribui%C3%A7%C3%A3o/?do=findComment&comment=364063
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@18622 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrTXT/ACBrLCDPR/Registro0040.pas
@@ -53,7 +53,7 @@ type
     FCOMPL: String;
     FCOD_MUN: String;
     FINSCR_ESTADUAL: String;
-    FPARTICIPACAO: Integer;
+    FPARTICIPACAO: Double;
     FENDERECO: String;
     FPAIS: String;
     FCAEPF: string;
@@ -69,7 +69,7 @@ type
     procedure SetNOME_IMOVEL(const Value: String);
     procedure SetNUM(const Value: String);
     procedure SetPAIS(const Value: String);
-    procedure SetPARTICIPACAO(const Value: Integer);
+    procedure SetPARTICIPACAO(const Value: Double);
     procedure SetTIPO_EXPLORACAO(const Value: TTipoExploracao);
     procedure SetUF(const Value: String);
     procedure SetCAEPF(const Value: string);
@@ -89,7 +89,7 @@ type
     property COD_MUN : String read FCOD_MUN write SetCOD_MUN;
     property CEP : String read FCEP write SetCEP;
     property TIPO_EXPLORACAO : TTipoExploracao read FTIPO_EXPLORACAO write SetTIPO_EXPLORACAO;
-    property PARTICIPACAO : Integer read FPARTICIPACAO write SetPARTICIPACAO;
+    property PARTICIPACAO : Double read FPARTICIPACAO write SetPARTICIPACAO;
   End;
 
 implementation
@@ -207,7 +207,7 @@ begin
   FPAIS := Value;
 end;
 
-procedure TRegistro0040.SetPARTICIPACAO(const Value: Integer);
+procedure TRegistro0040.SetPARTICIPACAO(const Value: Double);
 begin
   FPARTICIPACAO := Value;
 end;
